launcher: move launcher, pages, element and command into panel-plugin/launcher/
