+++ panel-plugin/launcher/applications-page.cpp
