+++ panel-plugin/launcher/applications-page.h
