+++ panel-plugin/launcher/command.cpp
