+++ panel-plugin/launcher/command.h
