+++ panel-plugin/launcher/element.cpp
