+++ panel-plugin/launcher/favorites-page.cpp
