+++ panel-plugin/launcher/favorites-page.h
