+++ panel-plugin/launcher/launcher.cpp
