+++ panel-plugin/launcher/launcher.h
