+++ panel-plugin/launcher/page.cpp
