+++ panel-plugin/launcher/page.h
