+++ panel-plugin/launcher/recent-page.cpp
