+++ panel-plugin/launcher/recent-page.h
