+++ panel-plugin/meson.build
@@ -1,13 +1,13 @@
 plugin_sources = [
-  'applications-page.cpp',
+  'launcher/applications-page.cpp',
   'presets/preset.cpp',
   'presets/preset-io.cpp',
   'launcher/category.cpp',
   'launcher/category-button.cpp',
-  'command.cpp',
+  'launcher/command.cpp',
   'command-edit.cpp',
-  'element.cpp',
-  'favorites-page.cpp',
+  'launcher/element.cpp',
+  'launcher/favorites-page.cpp',
   'places/favourites-section.cpp',
   'places/history-section.cpp',
   'places/home-search-worker.cpp',
@@ -15,17 +15,17 @@ plugin_sources = [
   'icon-renderer.cpp',
   'icon-size.cpp',
   'image-menu-item.h',
-  'launcher.cpp',
+  'launcher/launcher.cpp',
   'launcher-icon-view.cpp',
   'launcher-tree-view.cpp',
   'migration.cpp',
-  'page.cpp',
+  'launcher/page.cpp',
   'places/places-item.cpp',
   'places/places-page.cpp',
   'plugin.cpp',
   'profile.cpp',
   'search/query.cpp',
-  'recent-page.cpp',
+  'launcher/recent-page.cpp',
   'register-plugin.c',
   'resizer.cpp',
   'search/run-action.cpp',
