+++ panel-plugin/search/search-page.cpp
@@ -17,8 +17,8 @@
 
 #include "search-page.h"
 
-#include "applications-page.h"
-#include "launcher.h"
+#include "launcher/applications-page.h"
+#include "launcher/launcher.h"
 #include "launcher-view.h"
 #include "search-action.h"
 #include "settings.h"
