+++ po/POTFILES
@@ -1,10 +1,10 @@
-panel-plugin/applications-page.cpp
+panel-plugin/launcher/applications-page.cpp
 panel-plugin/launcher/category-button.cpp
 panel-plugin/launcher/category.cpp
 panel-plugin/command-edit.cpp
-panel-plugin/command.cpp
-panel-plugin/element.cpp
-panel-plugin/favorites-page.cpp
+panel-plugin/launcher/command.cpp
+panel-plugin/launcher/element.cpp
+panel-plugin/launcher/favorites-page.cpp
 panel-plugin/places/favourites-section.cpp
 panel-plugin/places/history-section.cpp
 panel-plugin/places/home-section.cpp
@@ -12,15 +12,15 @@ panel-plugin/icon-renderer.cpp
 panel-plugin/icon-size.cpp
 panel-plugin/launcher-icon-view.cpp
 panel-plugin/launcher-tree-view.cpp
-panel-plugin/launcher.cpp
-panel-plugin/page.cpp
+panel-plugin/launcher/launcher.cpp
+panel-plugin/launcher/page.cpp
 panel-plugin/places/places-item.cpp
 panel-plugin/places/places-page.cpp
 panel-plugin/plugin.cpp
 panel-plugin/preset.cpp
 panel-plugin/profile.cpp
 panel-plugin/query.cpp
-panel-plugin/recent-page.cpp
+panel-plugin/launcher/recent-page.cpp
 panel-plugin/resizer.cpp
 panel-plugin/run-action.cpp
 panel-plugin/search-action.cpp
